Add some javabeans :bulb:
I've added a new class VirtualMachine to account for the VMs that will
be created on this simulator, for dealocating purposes. They need to
register ID of itself and computer where they're located. They also
need to store the time that they will be dealocated for finishing its
usage. Finally the power of the VM, so that we know how much power will
be retrieved to computer.

I've also change the term processors for processor for better meaning.
And I've added some more functions to NodePower, so we can create it as
a JavaBeans too.