Add VMAllocator to use on simulator :bulb:
I've added a new class called VMAllocator. It's a Singleton responsible
to given a list of computers, and a selected algorithm, it allocates
VM desired to the computers. Also, whenever a allocation its required,
the allocator also search for VM that has reached its end time, to
deallocate from its computers.

Later, this class will collect metrics (fragmentation, and service
denial) on each allocation.

I've also added the IAllocatorAlgorithm interface. It's a Strategy with
the unique function of allocate a VM on a list of computers. We will be
able to select the algorithm used on our simulator, using this strategy
pattern.